+++ backend/src/test/kotlin/com/beatgear/backend/annotation/IntegrationTest.kt
@@ -1,9 +1,9 @@
 import io.restassured.RestAssured
+import org.junit.jupiter.api.AfterAll
+import org.junit.jupiter.api.BeforeAll
 import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.ActiveProfiles
-import org.springframework.test.context.TestContext
-import org.springframework.test.context.TestExecutionListener
-import org.springframework.test.context.TestExecutionListeners
+import org.springframework.test.context.*
+import org.testcontainers.containers.PostgreSQLContainer
 import org.testcontainers.junit.jupiter.Testcontainers
 
 @Target(AnnotationTarget.CLASS)
@@ -16,6 +16,37 @@ import org.testcontainers.junit.jupiter.Testcontainers
     mergeMode = TestExecutionListeners.MergeMode.MERGE_WITH_DEFAULTS
 )
 annotation class IntegrationTest {
+
+
+    companion object {
+        var postgres: PostgreSQLContainer<*> = PostgreSQLContainer(
+            "postgres:17-alpine"
+        )
+
+        @BeforeAll
+        @JvmStatic
+        internal fun beforeAll() {
+            postgres.start()
+        }
+
+        @AfterAll
+        @JvmStatic
+        internal fun afterAll() {
+            postgres.stop()
+        }
+
+        @DynamicPropertySource
+        @JvmStatic
+        internal fun configureProperties(registry: DynamicPropertyRegistry) {
+            registry.add("spring.datasource.url") { postgres.jdbcUrl }
+            registry.add("spring.datasource.username") { postgres.username }
+            registry.add("spring.datasource.password") { postgres.password }
+            registry.add("spring.flyway.url") { postgres.jdbcUrl }
+            registry.add("spring.flyway.user") { postgres.username }
+            registry.add("spring.flyway.password") { postgres.password }
+        }
+    }
+
     class ExecutionListener : TestExecutionListener {
         override fun beforeTestMethod(testContext: TestContext) {
             val applicationContext = testContext.applicationContext
